Move c return_types file
