+++ libvcx/src/api/disclosed_proof.rs
@@ -4,10 +4,10 @@ use indy_sys::CommandHandle;
 use libc::c_char;
 
 use crate::{connection, disclosed_proof};
-use crate::error::prelude::*;
 use crate::abi_utils::cstring::CStringUtils;
-use crate::utils::error;
 use crate::abi_utils::runtime::execute;
+use crate::error::prelude::*;
+use crate::utils::error;
 
 /*
     APIs in this module are called by a prover throughout the request-proof-and-verify process.
@@ -954,13 +954,14 @@ mod tests {
 
     use agency_client::mocking::AgencyMockDecrypted;
 
-    use crate::api::{return_types_u32, VcxStateType};
+    use crate::abi_utils::return_types_u32;
+    use crate::abi_utils::timeout::TimeoutUtils;
+    use crate::api::VcxStateType;
     use crate::utils::constants::{CREDS_FROM_PROOF_REQ, GET_MESSAGES_DECRYPTED_RESPONSE, V3_OBJECT_SERIALIZE_VERSION};
     use crate::utils::devsetup::*;
     use crate::utils::mockdata::mock_settings::MockBuilder;
     use crate::utils::mockdata::mockdata_credex::ARIES_CREDENTIAL_REQUEST;
     use crate::utils::mockdata::mockdata_proof::ARIES_PROOF_REQUEST_PRESENTATION;
-    use crate::abi_utils::timeout::TimeoutUtils;
 
     use super::*;
