+++ libvcx/src/api/filters.rs
@@ -64,7 +64,7 @@ mod tests {
 
     use crate::abi_utils::timeout::TimeoutUtils;
     use crate::api::filters::vcx_filter_proof_requests_by_name;
-    use crate::api::return_types_u32;
+    use crate::abi_utils::return_types_u32;
     use crate::connection;
     use crate::disclosed_proof::get_proof_request_messages;
     use crate::utils::{constants::GET_MESSAGES_DECRYPTED_RESPONSE, devsetup::*, error, mockdata::mockdata_proof};
