+++ libvcx/src/api/mod.rs
@@ -11,7 +11,6 @@ pub mod credential;
 pub mod disclosed_proof;
 pub mod wallet;
 pub mod logger;
-pub mod return_types_u32;
 mod filters;
 
 /// This macro allows the VcxStateType to be
